requirements.txt
fastapi
gunicorn
uvicorn
pyjwt
sqlalchemy
psycopg2-binary
celery
passlib
pandas
pycryptodome
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
slack-sdk
slack_bolt
atlassian-python-api
psycopg2
fastapi_keycloak
